requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.4
xgboost>=2.0
optuna>=3.6
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
openpyxl>=3.1
scipy>=1.11
jupyter>=1.0